test(#698): extract the real-hardware battery harness into a shared module
